requirement.txt
streamlit
langchain
langchain-groq
python-dotenv

